backend/requirements.txt
SD-Parsers
Pillow
PyExiv2
fastapi
uvicorn
python-multipart
